Features, Performance, and Readability
I added in the feature to select a maximum of two cards. Empty stars now display as filled ones are knocked off screen. Sound effects occur when a star is lost and when the volume is unmuted. I reduced interval called functions down to only one which is for the timer. Code reads cleaner and is better organized.